+++ my-app/App.tsx
@@ -1,24 +1,18 @@
-import { StyleSheet } from "react-native";
-import BottomNav from "./components/BottomNav";
-import { NavigationContainer } from "@react-navigation/native";
+import "react-native-gesture-handler";
 import { ThemeProvider } from "@rneui/themed";
 import { theme } from "./styles/theme";
+import Routes from "./Routes";
+import AuthProvider from "./Providers/AuthProvider";
+import UserProvider from "./Providers/UserProvider";
 
 export default function App() {
   return (
     <ThemeProvider theme={theme}>
-      <NavigationContainer>
-        <BottomNav />
-      </NavigationContainer>
+      <UserProvider>
+        <AuthProvider>
+          <Routes />
+        </AuthProvider>
+      </UserProvider>
     </ThemeProvider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
