+++ my-app/components/PageContainer/index.tsx
@@ -9,8 +9,6 @@ const PageContainer = ({ children }: { children: ReactNode }) => {
     <ScrollView
       style={{
         backgroundColor: theme.colors.white,
-        width: "100%",
-        height: "100%",
         padding: 5,
         paddingTop: 50,
       }}
