+++ my-app/screens/Private/HItchHike/index..tsx
@@ -1,8 +1,8 @@
 import { Text, View } from "react-native";
 import Stepper from "react-native-stepper-ui";
-import PageContainer from "../../components/PageContainer";
+import PageContainer from "../../../components/PageContainer";
 
-const MyComponent = (props:any) => {
+const MyComponent = (props: any) => {
   return (
     <View>
       <Text>{props.title}</Text>
@@ -11,13 +11,14 @@ const MyComponent = (props:any) => {
 };
 
 const content = [
-  <OriginAndDestination/>,
+  <OriginAndDestination />,
   <MyComponent title="Component 2" />,
   <MyComponent title="Component 3" />,
   <MyComponent title="Component 4" />,
 ];
-import { useState } from "react";
+import { useContext, useState } from "react";
 import OriginAndDestination from "./components/OriginAndDestination";
+import { userContext } from "../../../Providers/UserProvider";
 
 const HitchHike = () => {
   const [activeStep, setActiveStep] = useState(0);
@@ -26,7 +27,6 @@ const HitchHike = () => {
     <>
       <PageContainer>
         <Stepper
-          
           active={activeStep}
           content={content}
           onBack={() => setActiveStep((p) => p - 1)}
