+++ my-app/screens/Private/Profile/components/EditCarDialog/index.tsx
@@ -1,32 +1,31 @@
 import { Button, Dialog, useTheme, Input } from "@rneui/themed";
-import { ICar } from "../../../../interfaces/ICar";
+import { ICar } from "../../../../../interfaces/ICar";
 import { View } from "react-native";
-import Text from "../../../../components/Text";
+import Text from "../../../../../components/Text";
 
-interface IAddCarDialog {
+interface IEditCarDialog {
+  car: ICar;
   isVisible: boolean;
   onBackdropPress: () => void;
 }
 
-const AddCarDialog = ({ isVisible, onBackdropPress }: IAddCarDialog) => {
+const EditCarDialog = ({ car, isVisible, onBackdropPress }: IEditCarDialog) => {
   const { theme } = useTheme();
 
   return (
     <Dialog isVisible={isVisible} onBackdropPress={onBackdropPress}>
       <Text variant="Subtitle" style={{ textAlign: "center" }}>
-        Insira os dados do veículo
+        Edite o veículo
       </Text>
       <View style={{ marginTop: 20 }}>
-        <Input label="Modelo" />
-        <Input label="Placa" />
+        <Input label="Modelo" defaultValue={car.modelo} />
+        <Input label="Placa" defaultValue={car.placa} />
         <Button onPress={onBackdropPress} color="secondary" uppercase>
-          <Text variant="Action" style={{ color: theme.colors.white }}>
-            Confirmar
-          </Text>
+          <Text variant="Action" style={{color: theme.colors.white}}>Confirmar</Text>
         </Button>
       </View>
     </Dialog>
   );
 };
 
-export default AddCarDialog;
+export default EditCarDialog;
